move member to team members
