+++ packages/sdk/src/crud/user.ts
@@ -5,7 +5,7 @@ export interface User {
   email: string;
   metadata: {
     avatar_url: string;
-    full_name: string;
+    full_name?: string;
     username: string;
   };
 }
